pyproject.toml
[project]
name = "auto_psycho"
version = "0.1.0"
description = "Online Thematic Apperception Test (TAT) Experiment Platform with AI Analysis"
authors = [
    {name = "Sigurd", email = "sigurd@example.com"}
]
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "flask>=2.3.0",
    "flask-sqlalchemy>=3.0.0",
    "openai>=1.82.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pillow>=10.0.0",
    "jinja2>=3.1.0",
    "werkzeug>=2.3.0",
    "click>=8.1.0",
    "itsdangerous>=2.1.0",
    "markupsafe>=2.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_psycho"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 